+++ src/shared/particle_dynamics/fluid_dynamics/all_fluid_dynamics.h
@@ -40,3 +40,4 @@
 #include "fluid_dynamics_multi_phase.hpp"
 #include "fluid_surface_complex.h"
 #include "fluid_surface_inner.hpp"
+#include "transport_velocity_correction.h"
