+++ src/shared/particle_dynamics/fluid_dynamics/fluid_dynamics_complex.hpp
@@ -32,55 +32,6 @@ void DensitySummationComplexAdaptive::
                          sph_adaptation_.NumberDensityScaleFactor(h_ratio_[index_i]);
 }
 //=================================================================================================//
-void TransportVelocityCorrectionComplex::
-    interaction(size_t index_i, Real dt)
-{
-    TransportVelocityCorrectionInner::interaction(index_i, dt);
-
-    Vecd acceleration_trans = Vecd::Zero();
-    for (size_t k = 0; k < contact_configuration_.size(); ++k)
-    {
-        Neighborhood &contact_neighborhood = (*contact_configuration_[k])[index_i];
-        for (size_t n = 0; n != contact_neighborhood.current_size_; ++n)
-        {
-            Vecd nablaW_ijV_j = contact_neighborhood.dW_ijV_j_[n] * contact_neighborhood.e_ij_[n];
-
-            // acceleration for transport velocity
-            acceleration_trans -= 2.0 * nablaW_ijV_j;
-        }
-    }
-
-    /** correcting particle position */
-    if (surface_indicator_[index_i] == 0)
-        pos_[index_i] += coefficient_ * smoothing_length_sqr_ * acceleration_trans;
-}
-//=================================================================================================//
-void TransportVelocityCorrectionComplexAdaptive::
-    interaction(size_t index_i, Real dt)
-{
-    TransportVelocityCorrectionInnerAdaptive::interaction(index_i, dt);
-
-    Vecd acceleration_trans = Vecd::Zero();
-    for (size_t k = 0; k < contact_configuration_.size(); ++k)
-    {
-        Neighborhood &contact_neighborhood = (*contact_configuration_[k])[index_i];
-        for (size_t n = 0; n != contact_neighborhood.current_size_; ++n)
-        {
-            Vecd nablaW_ijV_j = contact_neighborhood.dW_ijV_j_[n] * contact_neighborhood.e_ij_[n];
-
-            // acceleration for transport velocity
-            acceleration_trans -= 2.0 * nablaW_ijV_j;
-        }
-    }
-
-    /** correcting particle position */
-    if (surface_indicator_[index_i] == 0)
-    {
-        Real inv_h_ratio = 1.0 / sph_adaptation_.SmoothingLengthRatio(index_i);
-        pos_[index_i] += coefficient_ * smoothing_length_sqr_ * inv_h_ratio * inv_h_ratio * acceleration_trans;
-    }
-}
-//=================================================================================================//
 void Oldroyd_BIntegration1stHalfWithWall::
     interaction(size_t index_i, Real dt)
 {
